--- apps/elf-eval/src/bin/real_world_job_benchmark/job_reports/local_organizer_reports.rs
@@ -1,1 +1,0 @@
-pub(crate) use crate::LocalOrganizerJobReport;
